requirements.txt
streamlit
langchain==0.1.14
faiss-cpu
python-dotenv
pypdf
tiktoken
google-generativeai
sentence-transformers==2.2.2
transformers==4.40.0